requirements.txt
pip==21.1.3
numpy==1.19.2
opencv-python==4.5.2.54
Pillow==8.2.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.1
opencv-python-headless==4.1.2.30
onnxruntime==1.13.1
insightface==0.7.2
